Improve Get Logs
I’m updating the Logs screen in the viam mobile app and noticed we’re missing many useful fields in our wrapper. exposing these fields, specifically start/end greatly improve the API call speed from taking ~45 seconds to ~.3 seconds.

I’ve also added an assert and test asserting endTime is after startTime. We do not enforce that on the backend so it can be confusing.

I’ve also updated the unit tests to include the new fields.

Also... this method does follow the cascading like the rest of the SDK now. I found using the cascade notation here with optional nullable fields introduced a lot of complexity needing to check if fields were null before setting them, because the setters from the generated proto code don’t accept nullable types.

I’m open to changing it to be like the rest of the SDK, but IMO this is more readable, and that would introduce a lot of one line if statements after creating `GetRobotPartLogsRequest()`.

Thanks!